+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Routes, Route, Navigate } from 'react-router-dom';
 
 import FeedDetailPage from '@/ui/pages/FeedDetailPage';
 
-import FeedList from '@/components/features/feed/FeedList';
+import FeedList from '@/components/features/feed/components/FeedList';
 import Login from '@/components/features/auth/Login';
 import SignUpForm from '@/components/features/auth/SignUpForm';
 import Profile from '@/components/features/profile/Profile';
@@ -19,7 +19,7 @@ import '@/App.css';
 
 import { ToastProvider } from '@/ui/components/common/ToastProvider';
 import SseListener from '@/ui/components/common/SseListener';
-import FeedAndMemberSearch from './components/features/feed/search/FeedAndMemberSearch';
+import FeedAndMemberSearch from './components/features/feed/search/Search';
 import Room from './components/features/dm/Room';
 
 function App() {
